ShouldThrowCriticalDependencyExceptionOnAddIfSqlErrorOccursAndLogItAsync -> PASS
