+++ Taarafo.Core/Services/Foundations/GroupPosts/GroupPostService.Exceptions.cs
@@ -5,6 +5,7 @@
 
 using System;
 using System.Threading.Tasks;
+using Microsoft.Data.SqlClient;
 using Taarafo.Core.Models.GroupPosts;
 using Taarafo.Core.Models.GroupPosts.Exceptions;
 using Xeptions;
@@ -29,6 +30,14 @@ namespace Taarafo.Core.Services.Foundations.GroupPosts
             {
                 throw CreateAndLogValidationException(invalidGroupPostException);
             }
+            catch(SqlException sqlException)
+            {
+                var failedGroupPostStorageException = 
+                    new FailedGroupPostStorageException(sqlException);
+
+                throw CreateAndLogCriticalDependencyException(
+                    failedGroupPostStorageException);
+            }
         }
 
         private GroupPostValidationException CreateAndLogValidationException(
@@ -40,6 +49,17 @@ namespace Taarafo.Core.Services.Foundations.GroupPosts
             this.loggingBroker.LogError(groupPostValidationException);
 
             return groupPostValidationException;
+        }
+
+        private GroupPostDependencyException CreateAndLogCriticalDependencyException(
+            Xeption exception)
+        {
+            var groupPostDependencyException =
+                new GroupPostDependencyException(exception);
+
+            this.loggingBroker.LogCritical(groupPostDependencyException);
+
+            return groupPostDependencyException;
         }
     }
 }
